Refactoring configui package
