+++ cmd/batocera/main.go
@@ -29,7 +29,7 @@ import (
 	"github.com/ZaparooProject/zaparoo-core/pkg/platforms/batocera"
 	"github.com/ZaparooProject/zaparoo-core/pkg/platforms/mister"
 	"github.com/ZaparooProject/zaparoo-core/pkg/service"
-	"github.com/ZaparooProject/zaparoo-core/pkg/ui"
+	"github.com/ZaparooProject/zaparoo-core/pkg/ui/tui"
 	"github.com/ZaparooProject/zaparoo-core/pkg/utils"
 	"github.com/rs/zerolog/log"
 	"os"
@@ -131,7 +131,7 @@ func main() {
 
 	// start the tui
 	logDestinationPath := path.Join(mister.DataDir, config.LogFile)
-	app, err := ui.BuildTheUi(pl, svc.Running(), cfg, logDestinationPath)
+	app, err := tui.BuildTheUi(pl, svc.Running(), cfg, logDestinationPath)
 	if err != nil {
 		log.Error().Msgf("error setting up UI: %s", err)
 		_, _ = fmt.Fprintf(os.Stderr, "Error setting up UI: %v\n", err)
