+++ pkg/ui/tui/buildapp_darwin.go
@@ -1,4 +1,4 @@
-package configui
+package tui
 
 import (
 	"github.com/rivo/tview"
